Add files via upload
I decided to estimate a unique \beta (which will be inflated or deflated in the simulation at individual level, depending on the risk perception) calibrated against the very first days of pandemic in Italy. The idea is that of estimating an evolution of the virus as it would have been in case no intervention had been made. So, I used as actual values the number of infected cases (without differentiating among those in isolation or hospitalized) up to March the 8th.
As data source, I used data collected, officially from February the 23rd, from the Italian Protezione Civile .