Add files via upload
In this notebook did some descriptive analysis of collected data. Then, I simulated the evolution of the virus with different versions of an Agent-Based Model with an underneath network structure. The diffusion mechanism I considered is a modified version of the compartmental S,I,R model.
The final plots show how modeling interventions leads to a reduction in the peak of the pandemic.